Add chameleonsocks-kolla-friendly.sh
This script should be run after a successful kolla deployment based
on HomeDC (such as kolla-prep-stein-bionic-source-*.sh) where a SOCKS
proxy is necessary to access the internet together with limited
external interfaces available for that purpose. In the original HomeDC
single-node deployment, there were two interfaces capable of reaching
the Internet, so there was complete separation between the interface
to be used as neutron_external_interface and the interface to be used
for the host itself to receive security updates, etc.

chameleonsocks-kolla-friendly.sh looks for the Neutron external router,
sets up chameleonsocks/redsocks in its network namespace, and configures
iptables so that guest VMs can have Internet out-of-the-box with zero
configuration needed.