Add fix-ext-router-ip-a-posteriori.sh
This script addresses a specific corner case in
one of HomeDC's deployments, where the single external
interface IP address can only be obtained by DHCP, even
though it's statically defined in the gateway interface of
the Neutron virtual router. The script takes care of
getting a new IP address and updating that in the respective
router's Neutron port, to keep everything consistent.