Add kolla-prep-one-master-bionic-source-v1.sh
This is the first kolla-prep-"one" script to deploy Kolla.
This variant is prepared to deal with the constraint of having a single
Internet-facing interface on the AIO host. A requirement is to have a
jumphost on the internal/mgmt/api network in order to access the machine
and the API.

Also switching to master this time, instead of stein, since stein:source
was no longer working due to a python36 incompatibility in Horizon at
the time I deployed it the first time in this "one"-constrained environment.